backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
aiofiles==23.2.1
pillow==10.1.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.4
torch==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2
openai-whisper==20231117
pydub==0.25.1
gtts==2.4.0
indicnlp==0.0.8
redis==5.0.1
celery==5.3.4
fastapi-limiter==0.1.6